from double to float

+++ includes/Connexion.h
@@ -4,10 +4,10 @@
 
 class Connexion {
 	public:
-		using Christoffel3D = std::array<std::array<std::array<double, NDIM>, NDIM>, NDIM>;
-		using Christoffel4D = std::array<std::array<std::array<std::array<double, NDIM>, NDIM>, NDIM>, NDIM>;
-		using MatrixNDIM = std::array<std::array<double, NDIM>, NDIM>;
-		using VectorNDIM = std::array<double, NDIM>;
+		using Christoffel3D = std::array<std::array<std::array<float, NDIM>, NDIM>, NDIM>;
+		using Christoffel4D = std::array<std::array<std::array<std::array<float, NDIM>, NDIM>, NDIM>, NDIM>;
+		using MatrixNDIM = std::array<std::array<float, NDIM>, NDIM>;
+		using VectorNDIM = std::array<float, NDIM>;
 
 		Christoffel3D Gamma{};
 		Christoffel4D Gamma_plus_h{};
@@ -15,10 +15,10 @@ class Connexion {
 		Christoffel4D Gamma_plus_half_h{};
 		Christoffel4D Gamma_minus_half_h{};
 
-		void calculate_christoffel(const VectorNDIM& X, double h,
+		void calculate_christoffel(const VectorNDIM& X, float h,
 				Christoffel3D& gamma,
-				std::array<std::array<double, NDIM>, NDIM>& g,
-				std::array<std::array<double, NDIM>, NDIM>& g_inv, 
+				std::array<std::array<float, NDIM>, NDIM>& g,
+				std::array<std::array<float, NDIM>, NDIM>& g_inv, 
 				const char* metric); 
 
 		void check_symmetry_christoffel(const Christoffel3D& gamma);
