+++ includes/Metric.h
@@ -4,14 +4,14 @@
 
 class Metric {
 	public:
-		std::array<std::array<double, NDIM>, NDIM> gcov{};
-		std::array<std::array<double, NDIM>, NDIM> gcon{};
-		std::array<std::array<double, NDIM>, NDIM> gcovK{};
-		std::array<std::array<double, NDIM>, NDIM> gconK{};
-		std::array<std::array<double, NDIM>, NDIM> gconMinkowski{};
-		std::array<std::array<double, NDIM>, NDIM> gcov_half{};
-		std::array<std::array<double, NDIM>, NDIM> gcon_half{};
-		std::array<std::array<double, NDIM>, NDIM> gcovMinkowski{};
+		std::array<std::array<float, NDIM>, NDIM> gcov{};
+		std::array<std::array<float, NDIM>, NDIM> gcon{};
+		std::array<std::array<float, NDIM>, NDIM> gcovK{};
+		std::array<std::array<float, NDIM>, NDIM> gconK{};
+		std::array<std::array<float, NDIM>, NDIM> gconMinkowski{};
+		std::array<std::array<float, NDIM>, NDIM> gcov_half{};
+		std::array<std::array<float, NDIM>, NDIM> gcon_half{};
+		std::array<std::array<float, NDIM>, NDIM> gcovMinkowski{};
 
 		Metric() = default;
 
@@ -20,16 +20,16 @@ class Metric {
 
 		
 		
-		void calculate_metric(const std::array<double, NDIM>& x, 
-				std::array<std::array<double, NDIM>, NDIM>& g,
-				std::array<std::array<double, NDIM>, NDIM>& g_inv);
-		void verify_metric(const std::array<std::array<double, NDIM>, NDIM>& g,
-				const std::array<std::array<double, NDIM>, NDIM>& g_inv);
-		void calculate_metric_kds(const std::array<double, NDIM>& x, 
-				std::array<std::array<double, NDIM>, NDIM>& g,
-				std::array<std::array<double, NDIM>, NDIM>& g_inv);
-		void calculate_metric_kerr_newman(const std::array<double, NDIM>& x, 
-				std::array<std::array<double, NDIM>, NDIM>& g,
-				std::array<std::array<double, NDIM>, NDIM>& g_inv);
+		void calculate_metric(const std::array<float, NDIM>& x, 
+				std::array<std::array<float, NDIM>, NDIM>& g,
+				std::array<std::array<float, NDIM>, NDIM>& g_inv);
+		void verify_metric(const std::array<std::array<float, NDIM>, NDIM>& g,
+				const std::array<std::array<float, NDIM>, NDIM>& g_inv);
+		void calculate_metric_kds(const std::array<float, NDIM>& x, 
+				std::array<std::array<float, NDIM>, NDIM>& g,
+				std::array<std::array<float, NDIM>, NDIM>& g_inv);
+		void calculate_metric_kerr_newman(const std::array<float, NDIM>& x, 
+				std::array<std::array<float, NDIM>, NDIM>& g,
+				std::array<std::array<float, NDIM>, NDIM>& g_inv);
 
 };
